requirements.txt
requests
rich
python-dotenv

